Convert tests to ReTestItems
