+++ Project.toml
@@ -5,8 +5,14 @@ version = "0.5.1"
 
 [deps]
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 MacroTools = "0.5"
 julia = "1.3"
+
+[extras]
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["ReTestItems", "Test"]
